fix: move components folder to presentation layer
