+++ library-build/entry.ts
@@ -1,3 +1,3 @@
-import StarryDigitizer from '../src/components/StarryDigitizer.vue'
+import StarryDigitizer from '../src/presentation/components/StarryDigitizer.vue'
 
 export { StarryDigitizer }
