+++ src/presentation/components/Canvas/CanvasAxesGuide.vue
