+++ src/presentation/components/Canvas/CanvasAxis.vue
