+++ src/presentation/components/Canvas/CanvasCursor.vue
