+++ src/presentation/components/Canvas/CanvasFooter.stories.js
