+++ src/presentation/components/Canvas/CanvasFooter.vue
