+++ src/presentation/components/Canvas/CanvasHeader.stories.js
