+++ src/presentation/components/Canvas/CanvasHeader.vue
