+++ src/presentation/components/Canvas/CanvasMain.vue
