+++ src/presentation/components/Canvas/CanvasPlot.vue
