+++ src/presentation/components/Export/Clipboard.stories.js
