+++ src/presentation/components/Generals/ConfirmerBar.vue
