+++ src/presentation/components/Magnifier/Magnifier.stories.js
