+++ src/presentation/components/Magnifier/MagnifierAxis.vue
