+++ src/presentation/components/Magnifier/MagnifierAxisLabelX.vue
