+++ src/presentation/components/Magnifier/MagnifierAxisLabelY.vue
