+++ src/presentation/components/Magnifier/MagnifierExtractSize.vue
