+++ src/presentation/components/Magnifier/MagnifierHorizontalLine.vue
