+++ src/presentation/components/Magnifier/MagnifierImage.vue
