+++ src/presentation/components/Magnifier/MagnifierMain.vue
@@ -59,7 +59,7 @@ import MagnifierAxes from './MagnifierAxes.vue'
 import MagnifierPlots from './MagnifierPlots.vue'
 import MagnifierSettings from './MagnifierSettings.vue'
 import MagnifierSettingsBtn from './MagnifierSettingsBtn.vue'
-import MagnifierExtractSize from '@/components/Magnifier/MagnifierExtractSize.vue'
+import MagnifierExtractSize from '@/presentation/components/Magnifier/MagnifierExtractSize.vue'
 import XYAxesCalculator from '@/domain/XYAxesCalculator'
 
 import { useAxesStore } from '@/store/axes'
