+++ src/presentation/components/Magnifier/MagnifierPlots.vue
