+++ src/presentation/components/Magnifier/MagnifierSettings.vue
