+++ src/presentation/components/Magnifier/MagnifierSettingsBtn.vue
