+++ src/presentation/components/Magnifier/MagnifierVerticalLine.vue
