+++ src/presentation/components/Magnifier/index.ts
