+++ src/presentation/components/Settings/AxesSettings.vue
