+++ src/presentation/components/Settings/ColorSettings.vue
