+++ src/presentation/components/Settings/ExtractorSettings.vue
