+++ src/presentation/components/Settings/ImageSettings.vue
