+++ src/presentation/components/Settings/LineExtractSettings.vue
