+++ src/presentation/components/Settings/SymbolExtractSettings.vue
