+++ src/presentation/components/StarryDigitizer.vue
@@ -33,13 +33,13 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import { MagnifierMain } from '@/components/Magnifier'
+import { MagnifierMain } from '@/presentation/components/Magnifier'
 import { CanvasHeader, CanvasFooter, CanvasMain } from './Canvas'
 import { AxesSettings, ExtractorSettings, ImageSettings } from './Settings'
 import { DatasetManager } from './DatasetManager'
-import { version } from '../../package.json'
-import ConfirmerBar from '@/components/Generals/ConfirmerBar.vue'
-import DataTable from '@/components/Export/DataTable.vue'
+import { version } from '../../../package.json'
+import ConfirmerBar from '@/presentation/components/Generals/ConfirmerBar.vue'
+import DataTable from '@/presentation/components/Export/DataTable.vue'
 
 export default defineComponent({
   components: {
@@ -120,3 +120,4 @@ $l_mainAreaSideMargin: 20px;
   }
 }
 </style>
+../../../package.json
